function addTogether is working like it should
Function addTogether detects if it has two parameters sent to it, then
tests if they are numbers, and if so, adds them together.  Else if it
has one numeric parameter sent in, it returns a function that takes one
numeric parameter and adds it to the original parameter sent in to its
parent function.