Added the tabulate method
The tabulate method takes a start and end parameters, and executes a
loop that uses said parameters, and executes the callback function on
each numeric value of i in the loop.
The callback function returns its value in each iteration of the loop
and passes it to the outArr array.