working solution found for independent button toggling
The color function is a closure that has access to the tog array.   The
color takes 2 parameters, the id and a num for the button being
manipulated.  If there is no such index as num, it makes one equal to 0.
It then sets the color of id as pink,  then increments the value of
tog[num].  the else conditional does the opposite.