Now btnVal wont  print out entire row of button at once
The btnVal function was printing out the entire row of the keyboard when
the row div was clicked, due to code using event.target to get the text
values of the buttons.
Now 'fixed' by a conditional that will not let it print out the string
if it is too long.